+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Updated the EC2 image ID to the latest Bottlerocket variant.
 - Added C8g instances to the `ClusterInstanceType` parameter values.
+- Changed CloudFront HTTP version to HTTP2/3.
 
 ### Fixed
 - Added `should-wait = true` to the EC2 instance configuration so it doesn't register itself in the ECS cluster while in warm pool.